feat: rename wacky-mixer files to mixer files
Rename all wacky-mixer tool files to match simplified naming:
- wacky-mixer-* → mixer-*

This aligns the code filenames with the simplified UI name "Mixer"
instead of "Electric Mixer".
