+++ js/tools/mixer-checkerboard.js
