+++ js/tools/mixer-pip.js
