+++ src/kidpix-main.js
@@ -65,13 +65,13 @@ import "../js/tools/stamp.js";
 import "../js/tools/three3d.js";
 import "../js/tools/tnt.js";
 import "../js/tools/trees.js";
-import "../js/tools/wacky-mixer-checkerboard.js";
-import "../js/tools/wacky-mixer-inverter.js";
-import "../js/tools/wacky-mixer-pattern.js";
-import "../js/tools/wacky-mixer-pip.js";
-import "../js/tools/wacky-mixer-shadow-boxes.js";
-import "../js/tools/wacky-mixer-venetian-blinds.js";
-import "../js/tools/wacky-mixer-wallpaper.js";
+import "../js/tools/mixer-checkerboard.js";
+import "../js/tools/mixer-inverter.js";
+import "../js/tools/mixer-pattern.js";
+import "../js/tools/mixer-pip.js";
+import "../js/tools/mixer-shadow-boxes.js";
+import "../js/tools/mixer-venetian-blinds.js";
+import "../js/tools/mixer-wallpaper.js";
 import "../js/tools/wholefx.js";
 
 // js/textures/*
